+++ src/Application/Cart/Read/Cart.php
@@ -22,13 +22,19 @@ interface Cart
     public function getQuantity(): int;
 
     public function getTotalPrice(?bool $includeTax = null): string;
-    public function getTotalPriceAsMoney(?bool $includeTax = null): Money;
     public function getSubtotalPrice(?bool $includeTax = null): string;
-    public function getSubtotalPriceAsMoney(?bool $includeTax = null): Money;
-    public function getShippingCost(): ?string;
-    public function getPaymentCost(): ?string;
-    public function getDiscountPrice(): ?string;
+    public function getShippingCost(?bool $includeTax = null): ?string;
+    public function getPaymentCost(?bool $includeTax = null): ?string;
+    public function getDiscountPrice(?bool $includeTax = null): ?string;
     public function getTaxPrice(): string;
+
+    public function getTotalPriceAsMoney(?bool $includeTax = null): Money;
+    public function getSubtotalPriceAsMoney(?bool $includeTax = null): Money;
+    public function getShippingCostAsMoney(?bool $includeTax = null): Money;
+    public function getPaymentCostAsMoney(?bool $includeTax = null): Money;
+    public function getDiscountPriceAsMoney(?bool $includeTax = null): Money;
+    public function getTaxPriceAsMoney(): Money;
+
     public function includeTax(bool $includeTax = true): void;
 
     public function getShopper(): ?CartShopper;